moved separated main from code
