+++ rpp.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <stdarg.h>
 
-#include "rpp.h"
+#include "preprocessor.h"
 
 int main(int argc, char **argv) {
     if (argc > 4 || argc == 1) {
@@ -24,19 +24,4 @@ int main(int argc, char **argv) {
         strcpy(outFileName, argv[2]);
     }
     preprocessRegFile(inFileName, outFileName);
-}
-
-int preprocessRegFile(char const *inFileName, char const *outFileName) {
-    FILE *inFile = fopen(inFileName, "r");
-    FILE *outFile = fopen(outFileName, "w");
-    doArithmeticPass(inFile, outFile);
-    fclose(inFile);
-    fclose(outFile);
-}
-
-int doArithmeticPass(FILE *inFile, FILE *outFile) {
-    char line[300] = {0};
-    while (fgets(line, sizeof(line), inFile)) {
-        
-    }
 }
